+++ lib/tag/endpoints.ts
@@ -5,9 +5,9 @@ import { endpoint } from "../endpoint";
 import { Prisma } from "@prisma/client";
 import { tagPolicy } from "./policy";
 import { Tag, TagExtended } from "./types";
-import { getTagAdapter } from "./service";
 import { UserAuth } from "../user/types";
-import { ResourceActions } from "../resource/types";
+import { RowActions } from "../common/types";
+import { TAG_NOT_EXISTS_ERROR } from "./constants";
 
 
 export const createTag = endpoint(async (
@@ -27,7 +27,7 @@ export const findTag = endpoint(async (
     args: Prisma.tagFindManyArgs
 ) => {
     return tagPolicy.filterForbidden(
-        ResourceActions.GET, await prisma.tag.findMany(args), user
+        RowActions.GET, await prisma.tag.findMany(args), user
     ) as TagExtended[];
 });
 
@@ -44,8 +44,10 @@ export const updateTag = endpoint(async (
     user: UserAuth,
     args: Prisma.tagUpdateArgs
 ) => {
-    const tag = await getTagAdapter({ where: args.where });
-    await tagPolicy.verifyAction(ResourceActions.UPDATE, tag, user);
+    const tag = await prisma.tag.findFirst({ where: args.where });
+    if (!tag) throw TAG_NOT_EXISTS_ERROR;
+
+    await tagPolicy.verifyAction(RowActions.UPDATE, tag, user);
     
     return prisma.tag.update(args);
 });
@@ -55,8 +57,10 @@ export const deleteTag = endpoint(async (
     user: UserAuth,
     args: Prisma.tagDeleteArgs
 ) => {
-    const tag = await getTagAdapter({ where: args.where });
-    await tagPolicy.verifyAction(ResourceActions.DELETE, tag, user);
+    const tag = await prisma.tag.findFirst({ where: args.where });
+    if (!tag) throw TAG_NOT_EXISTS_ERROR;
+
+    await tagPolicy.verifyAction(RowActions.DELETE, tag, user);
 
     return prisma.tag.delete(args);
 });
